% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allsinglearm.R
\name{allsinglearm}
\alias{allsinglearm}
\title{Onestage combination function for optimal designs}
\usage{
allsinglearm(p0, pa, alpha, beta, alpha2 = alpha, beta2, pi, eta, eps = 0.005)
}
\arguments{
\item{p0}{uninteresting response (null hypothesis H0), can be a vector}

\item{pa}{interesting response (alternative hypothesis Ha), can be a vector, always same length as p0. The corresponding elements of p0 and pa are taken as a set of proportions}

\item{alpha}{P(reject H0|H0) for Fleming and Simon designs, can be a vector}

\item{beta}{P(reject Ha|Ha)  for Fleming and Simon designs, can be a vector}

\item{alpha2}{P(reject H0|H0) for Sargent designs, can be a vector, same lenth as alpha}

\item{beta2}{P(reject Ha|Ha)  for Sargent designs, can be a vector, same length as beta}

\item{pi}{P(reject H0|Ha) for Sargent designs, can be a vector, same length as beta2}

\item{eta}{P(reject Ha|H0) for Sargent designs, can be a vector, same length as alpha2}

\item{eps}{tolerance (actual alpha<=alpha+eps; actual beta<=beta+eps; actual eta>=eta-eps; actual p>=pi-eps); default value = 0.005}
}
\description{
This function calculates sample sizes of the Fleming 1-stage , Simon-2stage, Sargent's 1- and 2-stage design. calculations are performed jointly for different sets of proportions (p0, pA) and operating characteristics. All designs are summarized in 2 tables, 1 for one-stage designs (Fleming and sargent) and 1 for two-stage designs (Simon and sargent)
}
\examples{
allsinglearm(p0 = 0.1, pa = 0.7,
             alpha = 0.05, beta = 0.2, beta2 = 0.1, pi = 0.8, eta = 0.8)
}
