% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app_start}
\alias{app_start}
\title{Launch Shiny app for Phase II clinical trials}
\usage{
app_start()
}
\description{
Web application allowing to see designs.
Note that this requires the rmarkdown, flexdashboard, shiny, shinyWidgets, ggplot2 and shinipsum packages to be installed
}
\examples{
\dontrun{
library(rmarkdown)
library(flexdashboard)
library(shiny)
library(shinyWidgets)
library(shinipsum)
library(ggplot2)
app_start()
}
}
