% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{data_sargent2}
\alias{data_sargent2}
\title{Dataset with design settings for sargent2stage}
\description{
Dataset as displayed in the following paper:
Sargent DJ, Goldberg RM. A Three-Outcome Design for Phase II Clinical Trials. Controlled Clinical Trials 22:117-125
}
\examples{
data(data_sargent2)
str(data_sargent2)
}
