% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleming1stage.R
\name{fleming1stage}
\alias{fleming1stage}
\title{The Fleming 1-stage function}
\usage{
fleming1stage(p0, pa, alpha, beta, eps = 0.005)
}
\arguments{
\item{p0}{uninteresting response (null hypothesis H0)}

\item{pa}{interesting response (alternative hypothesis Ha)}

\item{alpha}{P(reject H0|H0)}

\item{beta}{P(reject Ha|Ha)}

\item{eps}{tolerance (actual alpha<=alpha+eps; actual beta<=beta+eps; actual eta>=eta-eps; actual p>=pi-eps); default value = 0.005}
}
\description{
This function calculates sample sizes of the Fleming 1-stage design, for p0<pa
}
\details{
n= total number of patients

if x<=r --> futility
}
\examples{
fleming1stage(p0=0.45,pa=0.7,alpha=0.05,beta=0.2)
}
