% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleming2st_reject_Ha.R
\name{fleming2st_reject_Ha}
\alias{fleming2st_reject_Ha}
\title{P(reject Ha) in a 2-stage PhII single arm Fleming design}
\usage{
fleming2st_reject_Ha(n1, n2, r1, a, r2, p0, pa)
}
\arguments{
\item{n1}{total number of patients in stage1}

\item{n2}{total number of patients in stage2}

\item{r1}{critical value futility for the first stage}

\item{a}{critical value efficacy for the first stage}

\item{r2}{critical value futility/efficacy for the second stage}

\item{p0}{true success probability under H0}

\item{pa}{true success probability under Ha}
}
\description{
This function calculates the P(reject alternative hypothesis)
}
\details{
if x1<=r1 --> stop futility \cr
if x1>=a --> stop efficacy \cr
if (x1+x2)<=r2 --> futility \cr
with x1 the number of successes in the first stage and x2 the number of successes in the second stage
}
\examples{
fleming2st_reject_Ha(n1=13, n2=7, r1=0, a=3, r2=2, p0=0.05, pa=0.25)
}
\references{
Mander AP, Thompson SG. Two-stage designs optimal under the alternative hypothesis for phase II cancer clinical trials.
    Contemporary Clinical Trials 2010;31:572–578
    Qin F et al. Optimal, minimax and admissible two-stage design for phase II oncology clinical trials. BMC Medical Research Methodology 2020;20:126
}
