% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleming2stage_pval.R
\name{fleming2stage_pval}
\alias{fleming2stage_pval}
\title{p-value for a 2-stage PhII single arm Sargent design}
\usage{
fleming2stage_pval(n1, n2 = NULL, r1, a, k, p0)
}
\arguments{
\item{n1}{total number of patients in stage1}

\item{n2}{total number of patients in stage2}

\item{r1}{critical value futility for the first stage}

\item{a}{critical value efficacy for the first stage}

\item{k}{overall observed responses (must be larger than r1)}

\item{p0}{true success probability under H0}
}
\value{
p-value
}
\description{
This function calculates the p-value for a 2-stage PhII single arm Fleming's design
A stage-wise ordering of the sample space is used:
Let m be the stopping stage and x the total number of responses. A trial outcome (m'; x') is at least as
extreme (against H0) as the observed trial outcome (m; x), if one of the 3 following conditions is met:
(A) m' = m and x'>=x
(B) m' = 1, m = 2 and x'>=a
(C) m' = 2, m = 1 and x <=r1
in other words: rejection of H0 in the first stage is more extreme
}
\examples{
fleming2stage_pval(n1=21, n2=24, r1=5, a=10,k=17, p0=0.3)
}
\references{
Mander AP, Thompson SG. Two-stage designs optimal under the alternative hypothesis for phase II cancer clinical trials.
    Contemporary Clinical Trials 2010;31:572–578
    Qin F et al. Optimal, minimax and admissible two-stage design for phase II oncology clinical trials. BMC Medical Research Methodology 2020;20:126
    Nhacolo A, Brannath W. Interval and point estimation in adaptive Phase II trials with binary endpoint. Stat Methods Med Res 2019;28:2635-2648
}
