% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-2stage.R
\name{plot.2stage}
\alias{plot.2stage}
\title{Plot function for 2 stage sample size calculation}
\usage{
\method{plot}{`2stage`}(x, main = "Two-stage Designs", xlab = "Maximum Sample Size N", ylab, ...)
}
\arguments{
\item{x}{an object returned by \code{\link{simon2stage}} or \code{\link{sargent2stage}}}

\item{main}{title of the graph, passed on to \code{plot}}

\item{xlab}{x-axis label of the graph, passed on to \code{plot}}

\item{ylab}{y-axis label of the graph, passed on to \code{plot}}

\item{...}{other arguments passed on to \code{plot}}
}
\description{
Plots the expected sample size under H0 versus the maximum sample size
}
\examples{
samplesize <- simon2stage(p0 = 0.1, pa = 0.3, alpha = 0.05, beta = 0.2,
                          eps = 0.005, N_min = 0, N_max = 50)
plot(samplesize)

samplesize <- sargent2stage(p0 = 0.1, pa = 0.3, alpha = 0.05, beta = 0.1,
                            eta = 0.8, pi = 0.8,
                            eps = 0.005, N_min = 15, N_max = 30)
plot(samplesize)
}
