% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPhII.R
\name{plotPhII}
\alias{plotPhII}
\title{Plot function for 1- and 2-stage Sargent design}
\usage{
plotPhII(
  x,
  design_nr = 1,
  main = "PhII design",
  xlab = "True response probability",
  ylab,
  ...
)
}
\arguments{
\item{x}{an object returned by \code{\link{sargent1stage}} or \code{\link{sargent2stage}}}

\item{design_nr}{design_nr returned by \code{\link{sargent1stage}} or \code{\link{sargent2stage}} \cr
default is 1 (for \code{\link{sargent1stage}}, only one design is outputted)}

\item{main}{title of the graph, passed on to \code{plot}}

\item{xlab}{x-axis label of the graph, passed on to \code{plot}}

\item{ylab}{y-axis label of the graph, passed on to \code{plot}}

\item{...}{other arguments passed on to \code{plot}}
}
\description{
Plots the probability of rejecting \eqn{H0} ('Positive result') and rejecting Ha \cr
('Negative result') for different true response binomial parameters \cr
if the input is an object returned by \code{\link{sargent2stage}}, interactively choose \cr
the design number
}
\examples{
\donttest{
fleming1 <- fleming1stage(p0 = 0.45, pa = 0.7, alpha = 0.05, beta = 0.2)
plotPhII(fleming1)
simon2<- simon2stage(p0 = 0.1, pa = 0.3, alpha = 0.05, beta = 0.2,
                     eps = 0.005, N_min = 1, N_max = 50)
plotPhII(simon2)

sargent1 <- sargent1stage(p0 = 0.2, pa = 0.35, alpha = 0.1, beta = 0.1, eta = 0.8, pi = 0.8,
                         eps = 0.005, N_min = 35, N_max = 50)
plotPhII(sargent1)

sargent2 <- sargent2stage(p0 = 0.1, pa = 0.3, alpha = 0.05, beta = 0.1,
                          eta = 0.8, pi = 0.8,
                          eps = 0.005, N_min = 15, N_max = 30)
plotPhII(sargent2,design_nr=1)
}
}
\references{
Sargent DJ, Chan V, Goldberg RM. A three-outcome design for phase II clinical \cr
trials. Control Clin Trials. 2001;22(2):117-125. doi:10.1016/s0197-2456(00)00115-x \cr
Simon R. Optimal two-Stage Designs for Phase II Clinical Trials. \cr
Control Clin Trials. 1989;10:1-10
}
