% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSargent.R
\name{PlotSargent}
\alias{PlotSargent}
\title{Plot function for 1 stage Sargent design}
\usage{
PlotSargent(
  x,
  main = "Sargent's 3-outcome 1-stage design",
  xlab = "True response probability",
  ylab,
  ...
)
}
\arguments{
\item{x}{an object returned by \code{\link{sargent1stage}}}

\item{main}{title of the graph, passed on to \code{plot}}

\item{xlab}{x-axis label of the graph, passed on to \code{plot}}

\item{ylab}{y-axis label of the graph, passed on to \code{plot}}

\item{...}{other arguments passed on to \code{plot}}
}
\description{
Plots the probability of rejecting \eqn{H0} ('Positive result') and rejecting Ha \cr
('Negative result') for different true response binomial parameters
}
\examples{
result<-sargent1stage(p0 = 0.2, pa = 0.35, alpha = 0.1, beta = 0.1, eta = 0.8, pi = 0.8,
                      eps = 0.005, N_min = 35, N_max = 50)
PlotSargent(result)
}
