% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sargent1stage.R
\name{sargent1stage}
\alias{sargent1stage}
\title{The Sargent 1-stage function}
\usage{
sargent1stage(p0, pa, alpha, beta, eta, pi, eps = 0.005, N_min, N_max)
}
\arguments{
\item{p0}{uninteresting response (null hypothesis H0)}

\item{pa}{interesting response (alternative hypothesis Ha)}

\item{alpha}{P(reject H0|H0)}

\item{beta}{P(reject Ha|Ha)}

\item{eta}{P(reject Ha|H0)}

\item{pi}{P(reject H0|Ha)}

\item{eps}{tolerance (actual alpha<=alpha+eps; actual beta<=beta+eps; actual eta>=eta-eps; actual p>=pi-eps); default value = 0.005}

\item{N_min}{minimum sample size value for grid search}

\item{N_max}{maximum sample size value for grid search}
}
\description{
This function calculates sample sizes of the sargent 1-stage design.
}
\details{
if x<=r --> stop futility

if x>=s --> efficacy
}
\examples{
sargent1stage(p0=0.5,pa=0.65,alpha=0.1,beta=0.1,eta=0.8,pi=0.8,eps = 0.005,N_min=0,N_max=100)
}
