% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sargent2stage.R
\name{sargent2stage}
\alias{sargent2stage}
\title{The Sargent 2-stage function}
\usage{
sargent2stage(p0, pa, alpha, beta, eta, pi, eps = 0.005, N_min, N_max)
}
\arguments{
\item{p0}{uninteresting response (null hypothesis H0)}

\item{pa}{interesting response (alternative hypothesis Ha)}

\item{alpha}{P(reject H0|H0)}

\item{beta}{P(reject Ha|Ha)}

\item{eta}{P(reject Ha|H0)}

\item{pi}{P(reject H0|Ha)}

\item{eps}{tolerance (actual alpha<=alpha+eps; actual beta<=beta+eps; actual eta>=eta-eps; actual p>=pi-eps); default value = 0.005}

\item{N_min}{minimum sample size value for grid search}

\item{N_max}{maximum sample size value for grid search}
}
\description{
This function calculates sample sizes of the sargent 2-stage design.
}
\details{
n1= n stage1

n2= n stage2

total number of patients=n1+n2

if x1<=r1 --> stop futility

if (x1+x2)<=r --> futility

if (x1+x2)>=s --> efficacy
}
\examples{
sargent2stage(p0=0.1,pa=0.3,alpha=0.05,beta=0.1,eta=0.8,pi=0.8,eps = 0.005,N_min=15,N_max=30)
}
