% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sargent2stage_reject_Ha.R
\name{sargent2stage_prob_reject_Ha}
\alias{sargent2stage_prob_reject_Ha}
\title{P(reject Ha) in a 2-stage PhII single arm Sargent design}
\usage{
sargent2stage_prob_reject_Ha(n1, n2, r1, r2, p)
}
\arguments{
\item{n1}{total number of patients in stage1}

\item{n2}{total number of patients in stage2}

\item{r1}{critical value for the first stage}

\item{r2}{critical value for the second stage}

\item{p}{true success probability}
}
\value{
Probability
}
\description{
This function calculates the P(reject alternative hypothesis|true success probability)
}
\details{
if x1<=r1 --> stop futility \cr
if (x1+x2)<=r --> futility \cr
with x1 the number of successes in the first stage and x2 the number of successes in the second stage
}
\examples{
sargent2stage_prob_reject_Ha(n1=10,n2=20,r1=0,r2=4,p=0.25)
}
\references{
Sargent DJ, Chan V, Goldberg RM. A three-outcome design for phase II clinical trials. Control Clin Trials. 2001;22(2):117-125. doi:10.1016/s0197-2456(00)00115-x
}
