% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sargent2stage_pval.R
\name{sargent2stage_pval}
\alias{sargent2stage_pval}
\title{p-value for a 2-stage PhII single arm Sargent design}
\usage{
sargent2stage_pval(n1, n2 = NULL, r1, k, p0)
}
\arguments{
\item{n1}{total number of patients in stage1}

\item{n2}{total number of patients in stage2}

\item{r1}{critical value for the first stage}

\item{k}{overall observed responses (must be larger than r1)}

\item{p0}{true success probability under H0}
}
\value{
p-value
}
\description{
This function calculates the p-value for a 2-stage PhII single arm Sargent design
In the case of Sargent's 2-stage design, stage-wise ordering means that outcomes observed in
the second stage of the trial are more extreme than outcomes observed in the first stage of the trial.
}
\examples{
sargent2stage_pval(n1=15, n2=10, r1=1, k=6, p0=0.1)
}
\references{
Simon R. Optimal two-Stage Designs for Phase II Clinical Trials. \cr
Control Clin Trials. 1989;10:1-10; Koyama T and Chen H (2008): Proper inference from Simon's two-stage designs. \cr
Statistics in Medicine, 27(16):3145-3154.
}
