% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simon2stage_reject_Ha.R
\name{simon2stage_prob_reject_Ha}
\alias{simon2stage_prob_reject_Ha}
\title{P(reject Ha) in a 2-stage PhII single arm Simon design}
\usage{
simon2stage_prob_reject_Ha(n1, n2, r1, r, p)
}
\arguments{
\item{n1}{total number of patients in stage1}

\item{n2}{total number of patients in stage2}

\item{r1}{critical value for the first stage}

\item{r}{critical value for the second stage}

\item{p}{true success probability}
}
\value{
Probability
}
\description{
This function calculates the P(reject alternative hypothesis|true success probability)
}
\details{
if x1<=r1 --> stop futility \cr
if (x1+x2)<=r --> futility \cr
with x1 the number of successes in the first stage and x2 the number of successes in the second stage
}
\examples{
simon2stage_prob_reject_Ha(n1=10,n2=20,r1=0,r=4,p=0.25)
}
\references{
Simon R. Optimal two-Stage Designs for Phase II Clinical Trials. \cr
Control Clin Trials. 1989;10:1-10
}
