% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_1stage_R.R
\name{exact1stage}
\alias{exact1stage}
\title{The exact 1-stage function, randomized}
\usage{
exact1stage(p0, pa, alpha, beta, eps = 0.005, alloc, ...)
}
\arguments{
\item{p0}{uninteresting response (null hypothesis H0)}

\item{pa}{interesting response (alternative hypothesis Ha)}

\item{alpha}{P(reject H0|H0)}

\item{beta}{P(reject Ha|Ha)}

\item{eps}{tolerance (actual alpha<=alpha+eps; actual beta<=beta+eps; actual eta>=eta-eps; actual p>=pi-eps); default value = 0.005}

\item{alloc}{allocation ratio (e.g. 2 for 2:1)}

\item{...}{refers to type="exact" or "normal" for pdf and cdf difference of two binomial variables}
}
\description{
This function calculates sample sizes of an exact randomized 1-stage design, for p0<pa
}
\details{
n= total number of patients

n_E= number of patients in experimental arm

n_C= number of patients in control arm

if (E-C)<=r --> futility

if (E-C)> r --> efficacy

Note that all sample sizes can be entered, but only calculations for sample sizes, without any decimals (taking into account allocation ratio)

Variable of interest: diff=E-C

alloc is assumed >=1 (more patients in the experimental arm)
}
\examples{
exact1stage(p0 = 0.45, pa = 0.7, alpha = 0.05, beta = 0.2, alloc = 1, type = "normal")
}
