% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_1stage_R.R
\name{exact1stage}
\alias{exact1stage}
\title{The exact 1-stage function, randomized}
\usage{
exact1stage(p0, pa, alpha, beta, eps = 0.005, alloc = 1, type = "exact")
}
\arguments{
\item{p0}{probability of the uninteresting response (null hypothesis H0)}

\item{pa}{probability of the interesting response (alternative hypothesis Ha)}

\item{alpha}{Type I error rate \eqn{P(reject H0|H0)}}

\item{beta}{Type II error rate \eqn{P(reject Ha|Ha)}}

\item{eps}{tolerance default value = 0.005}

\item{alloc}{allocation ratio (e.g. 2 for 2:1). alloc is assumed >=1 (more patients in the experimental arm)}

\item{type}{character string, either "exact" or "normal" for pdf and cdf difference of two binomial variables}
}
\value{
a data.frame with elements
\itemize{
\item n: total number of patients
\item n_E: number of patients in experimental arm
\item n_C: number of patients in control arm
\item r: TODO. Variable of interest: diff=E-C. Note if (E-C)<=r --> futility\cr
\item d: TODO. Variable of interest: diff=E-C. Note if if (E-C)> r --> efficacy
\item alpha: the actual alpha value which is smaller than \code{alpha_param + eps}
\item beta: the actual beta value where which is smaller than \code{beta_param + eps}
\item p0: your provided \code{p0} value
\item pa: your provided \code{pa} value
\item alpha_param: your provided \code{alpha} value
\item beta_param: your provided \code{beta} value
}
}
\description{
This function calculates sample sizes of an exact randomized 1-stage design, for p0<pa
}
\details{
Note that all sample sizes can be entered, but only calculations for sample sizes, without any decimals (taking into account allocation ratio)
}
\examples{
exact1stage(p0 = 0.45, pa = 0.7, alpha = 0.05, beta = 0.2, alloc = 1, type = "normal")
exact1stage(p0 = 0.45, pa = 0.7, alpha = 0.05, beta = 0.2, alloc = 1, type = "exact")

test <- rbind(
  data.frame(p0 = c(0.1, 0.2, 0.3, 0.4),
             pa = c(0.1, 0.2, 0.3, 0.4) + 0.2),
  data.frame(p0 = c(0.1, 0.2, 0.3, 0.4),
             pa = c(0.1, 0.2, 0.3, 0.4) + 0.15))
test <- merge(test,
              expand.grid(alpha = c(0.05, 0.1), beta = 0.1))
samplesize <- exact1stage(p0 = test$p0, pa = test$pa,
                          alpha = test$alpha, beta = test$beta,
                          eps = 0.005, alloc = 1, type = "normal")
samplesize <- exact1stage(p0 = test$p0, pa = test$pa,
                          alpha = test$alpha, beta = test$beta,
                          eps = 0.2, alloc = 1, type = "exact")
## these 2 are the same
samplesize <- exact1stage(p0 = test$p0, pa = test$pa,
                          alpha = test$alpha, beta = test$beta,
                          eps = 0.005, alloc = 1, type = "normal")
\donttest{
samplesize <- mapply(p0 = test$p0, pa = test$pa, alpha = test$alpha, beta = test$beta,
  FUN=function(p0, pa, alpha, beta){
    exact1stage(p0 = p0, pa = pa, alpha = alpha, beta = beta,
                eps = 0.005, alloc = 1, type = "normal")
  }, SIMPLIFY = FALSE)
samplesize <- do.call(rbind, samplesize)
}
}
