% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCI_Koyama.R
\name{getCI_Koyama}
\alias{getCI_Koyama}
\title{Koyama confidence interval.}
\usage{
getCI_Koyama(k, r1, n1, n, alpha = 0.05, precision = 4)
}
\arguments{
\item{k}{overall observed responses (must be larger than r1).}

\item{r1}{critical value for the first stage.}

\item{n1}{sample size for the first stage.}

\item{n}{overall sample size.}

\item{alpha}{determining the two sided 1-2*alpha confidence interval.}

\item{precision}{gives the precision (in decimal numbers) to which the confidence interval should be calculated (should be less than 10).}
}
\description{
Calculates the two sided 1-2*alpha confidence interval based on the work from Koyama and Chen.
}
\examples{
#Calculate a Simon's two-stage design
res <- simon2stage(p0 = 0.1, pa = 0.3, alpha = 0.05, beta = 0.2,eps = 0.005, N_min = 1,
N_max = 50)
CI <- mapply(a = res$r2 + 1, b = res$r1, c = res$n1, d = res$N, FUN = function(a, b, c, d)
getCI_Koyama(k = a, r1 = b, n1 = c, n = d, alpha = 0.05, precision = 3))
}
\references{
Koyama T and Chen H (2008): Proper inference from Simon's two-stage designs. Statistics in Medicine, 27(16):3145-3154.
}
