% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSargent.R
\name{plotsargent}
\alias{plotsargent}
\title{Plot function for 1- and 2-stage Sargent design}
\usage{
plotsargent(
  x,
  design_nr = 1,
  main = "Sargent's 3-outcome 1-stage design",
  xlab = "True response probability",
  ylab,
  ...
)
}
\arguments{
\item{x}{an object returned by \code{\link{sargent1stage}} or \code{\link{sargent2stage}}}

\item{design_nr}{design_nr returned by \code{\link{sargent1stage}} or \code{\link{sargent2stage}} \cr
default is 1 (for \code{\link{sargent1stage}}, only one design is outputted)}

\item{main}{title of the graph, passed on to \code{plot}}

\item{xlab}{x-axis label of the graph, passed on to \code{plot}}

\item{ylab}{y-axis label of the graph, passed on to \code{plot}}

\item{...}{other arguments passed on to \code{plot}}
}
\description{
Plots the probability of rejecting \eqn{H0} ('Positive result') and rejecting Ha \cr
('Negative result') for different true response binomial parameters \cr
if the input is an object returned by \code{\link{sargent2stage}}, interactively choose \cr
the design number
}
\examples{
\donttest{
result_1<- sargent1stage(p0 = 0.2, pa = 0.35, alpha = 0.1, beta = 0.1, eta = 0.8, pi = 0.8,
                         eps = 0.005, N_min = 35, N_max = 50)
plotsargent(result_1)

result_2 <- sargent2stage(p0 = 0.1, pa = 0.3, alpha = 0.05, beta = 0.1,
                          eta = 0.8, pi = 0.8,
                          eps = 0.005, N_min = 15, N_max = 30)
plotsargent(result_2)
}
}
