% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sargent2stage.R
\name{sargent2stage}
\alias{sargent2stage}
\title{The Sargent 2-stage function}
\usage{
sargent2stage(
  p0,
  pa,
  alpha,
  beta,
  eta,
  pi,
  eps = 0,
  N_min,
  N_max,
  int = 0,
  int_window = 0.025
)
}
\arguments{
\item{p0}{probability of the uninteresting response (null hypothesis \eqn{H0})}

\item{pa}{probability of the interesting response (alternative hypothesis Ha)}

\item{alpha}{Type I error rate \eqn{P(reject H0|H0)}}

\item{beta}{Type II error rate \eqn{P(reject Ha|Ha)}}

\item{eta}{\eqn{P(reject Ha|H0)}}

\item{pi}{\eqn{P(reject H0|Ha)}}

\item{eps}{tolerance default value = 0.005}

\item{N_min}{minimum sample size value for grid search}

\item{N_max}{maximum sample size value for grid search}

\item{int}{pre-specified interim analysis percentage information}

\item{int_window}{window around interim analysis percentage (e.g. 0.5 +- 0.025). 0.025 is default value}
}
\value{
a data.frame with elements
\itemize{
\item n1: total number of patients in stage1
\item n2: total number of patients in stage2
\item N: total number of patients=n1+n2
\item r1: critical value for the first stage
\item r2: critical value for the second stage
\item eff: s/N
\item 90%CI_low: Result of call to getCI_Koyama. Confidence interval according to Koyama and Chen (1989)
\item 90%CI_high: Result of call to getCI_Koyama. Confidence interval according to Koyama and Chen (1989)
\item EN.p0: expected sample size under H0
\item PET.p0: probability of terminating the trial at the end of the first stage under H0
\item MIN: column indicating if the design is the minimal design
\item OPT: column indicating if the setting is the optimal design
\item ADMISS: column indicating if the setting is the admissible design
\item alpha: the actual alpha value which is smaller than \code{alpha_param + eps}
\item beta: the actual beta value where which is smaller than \code{beta_param + eps}
\item eta: the actual eta value which is smaller than \code{eta_param - eps}
\item pi: the actual pi value which is smaller than \code{pi_param - eps}
\item lambda:  1-(eta+alpha)
\item delta: 1-(beta+pi)
\item p0: your provided \code{p0} value
\item pa: your provided \code{pa} value
\item alpha_param: your provided \code{alpha} value
\item beta_param: your provided \code{beta} value
\item eta_param: your provided \code{eta} value
\item pi_param: your provided \code{pi} value
}
}
\description{
This function calculates sample sizes of the Sargent 2-stage design.

The goal of a phase II trial is to make a preliminary determination regarding the activity and
tolerability of a new treatment and thus to determine whether the treatment warrants
further study in the phase III setting. \cr
This function calculates the sample size needed in a Sargent 2-stage design which is a
three-outcome design that allows for three outcomes: reject \eqn{H(0)}, reject \eqn{H(a)}, or reject neither.
}
\details{
if x1<=r1 --> stop futility \cr
if (x1+x2)<=r --> futility \cr
if (x1+x2)>=s --> efficacy \cr
}
\examples{
samplesize <- sargent2stage(p0 = 0.1, pa = 0.3, alpha = 0.05, beta = 0.1, eta = 0.8, pi = 0.8,
                            eps = 0.005, N_min = 15, N_max = 30)
plot(samplesize)

\donttest{
data(data_sargent2)
test <- data_sargent2
samplesize <- sargent2stage(p0 = test$p0, pa = test$pa, alpha = test$alpha, beta = test$beta,
                            eta = test$eta, pi = test$pi,
                            eps = 0.005,
                            N_min = test$N_min, N_max = test$N_max)
optimal <- lapply(samplesize, FUN=function(x) subset(x, OPT == "Optimal"))
optimal <- data.table::rbindlist(optimal)
minimax <- lapply(samplesize, FUN=function(x) subset(x, MIN == "Minimax"))
minimax <- data.table::rbindlist(minimax)
}
}
\references{
Sargent DJ, Chan V, Goldberg RM. A three-outcome design for phase II clinical trials. Control Clin Trials. 2001;22(2):117-125. doi:10.1016/s0197-2456(00)00115-x
}
